Client sends add and delete user in organisation request

+++ src/pages/api/[orgKey]/users.ts
@@ -2,13 +2,17 @@ import { NextApiRequest, NextApiResponse } from "next";
 import { getQueryParam, getRequestBody } from "@/server/utils";
 import OrganisationRepository from "@/server/dao/OrganisationRepository";
 import UserRepository from "@/server/dao/UserRepository";
+import { getUserToken } from "@/server/auth/session";
+import { JWT } from "next-auth/jwt";
 
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ): Promise<void> {
   const organisation: string = getQueryParam(req, "orgKey");
-  const userEmail: string = getRequestBody(req, "newuser");
+  const userEmail: string = getRequestBody(req, "email");
+  const token: JWT = await getUserToken(req);
+  const ownEmail = token.email;
 
   if (organisation === "") {
     res.status(405).end();
@@ -19,7 +23,6 @@ export default async function handler(
     switch (req.method) {
       case "GET":
         const users = await UserRepository.getAllUsers({ organisation });
-        console.log(JSON.stringify(users));
         res.setHeader("Content-Type", "application/json");
         res.status(200).end(JSON.stringify(users));
         break;
@@ -37,6 +40,13 @@ export default async function handler(
         }
         break;
       case "DELETE":
+        if (userEmail === ownEmail) {
+          res.status(406).send({
+            message: "User cannot remove themselves.",
+          });
+          return;
+        }
+
         await OrganisationRepository.removeUser(userEmail, organisation);
         res.status(200).end();
         break;
